requirements.txt
Flask
Flask_Cors
docker
pydantic
neo4j
requests
python-dotenv
sentence-transformers